Architecture refactored
