+++ MarkersDetector.py
@@ -4,20 +4,14 @@ from parameters import ScreenPosition
 from typing import List
 
 
-class DetectedMarkersStatus:
-    """
-    Contains data to describe a list of several markers
-    """
-    corners: List[ScreenPosition] = []
-    ids: List = []
-
-
 class MarkersDetector:
     """
     Detects every marker on the frame coming from the Tello front camera,
     then returns a single DetectedMarkersStatus class containing data for all detected markers
     """
     PARAM_DRAW_MARKERS: bool = True
+    corners: List[ScreenPosition] = []
+    ids: List = []
 
     @classmethod
     def run(cls, frame: numpy.ndarray) -> numpy.ndarray:
@@ -25,8 +19,8 @@ class MarkersDetector:
         corners, ids = cls.find_markers(cp_frame)
         if cls.PARAM_DRAW_MARKERS and ids is not None:
             cls.draw_markers(cp_frame, corners, ids)
-        DetectedMarkersStatus.ids = ids
-        DetectedMarkersStatus.corners = corners
+        cls.ids = ids
+        cls.corners = corners
         return cp_frame
 
     @classmethod
